Moved useSpringTransition to a hook directory
